update: fix async-app app chaining when mounting sub app
When mounting a sub app with `app.use(path, subApp)`, express does a
check to see if an app was sent through, or if it was a middleware.

In async-app, we are converting all MiddlewareArguments which confuses
express to think that the mounted app isn't an sub application, but a
simple middleware.

Thus, express doesn't fire the `mount` event, which in turns skips a
bunch of express routines:

 - sets the `mountpath` on the subApp
 - copy some settings from the parent app (for instance "trust proxy")

With this change, we can now use `req.app.path` to get the full route
the request was matched for, and `req.ip` will be properly propagated
through.